refactor: useDynamicVisibleCount 공용 훅으로 이동
